implement a workaround for Java 11 (#10)
Due to a bug, Eclipse Java Compiler is not able to read in memory files,
yet it looks straight on the permanent file system.

See https://bugs.eclipse.org/bugs/show_bug.cgi?id=541269

Hence, this revision creates a temporary version of the file on
the permanent file system in order to make EJC happy.